Все console вызовы удалены
