+++ src/contexts/contexts.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+// Создаем контекст задач
+export const TasksContext = createContext(null);
